email-tracker-flask/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.2.0
geoip2==4.7.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms[email]==3.1.2
Flask-Migrate==4.0.5
Werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0 # <-- Add Gunicorn